diagnostic code

+++ qmcpy/discrete_distribution/halton/halton_owen.py
@@ -40,7 +40,7 @@ class HaltonOwen(object):
         return x
     
     def set_seed(self, seed):
-        self.s = seed if seed else random.randint(2**32)
+        self.s = seed if seed else random.randint(2**32, dtype=uint64)
         return self.s
         
     def set_dimension(self, dimension):
